[GLUTEN-1209][VL]refactor: Refactor Java Celeborn into an independent module
