+++ gluten-data/pom.xml
@@ -42,10 +42,6 @@
 
   <dependencies>
     <dependency>
-      <groupId>org.apache.celeborn</groupId>
-      <artifactId>celeborn-client-spark-${spark.major.version}-shaded_${scala.binary.version}</artifactId>
-    </dependency>
-    <dependency>
       <groupId>io.glutenproject</groupId>
       <artifactId>gluten-core</artifactId>
       <version>${project.version}</version>
